Language.Haskell.DoNotation

Description

This module provides new implementations for (>>=), (>>), pure and return so that they will work simultaneously with both regular and indexed monads.

Intended usage:

@@ {-# LANGUAGE RebindableSyntax #-}

import Language.Haskell.DoNotation import Prelude hiding (Monad (..), pure) @@

Documentation

class BindSyntax (x :: Type -> Type) (y :: Type -> Type) (z :: Type -> Type) | x y -> z, x z -> y, y z -> x where

Typeclass that provides (>>=) and (>>).

Minimal complete definition

(>>=)

Methods

(>>=) :: x a -> (a -> y b) -> z b

(>>) :: x a -> y b -> z b

class PureSyntax (x :: Type -> Type) where

Typeclass that provides pure and return.

Minimal complete definition

Nothing

Methods

pure :: a -> x a

return :: a -> x a

class Applicative m => Monad (m :: Type -> Type)

The Monad class defines the basic operations over a monad, a concept from a branch of mathematics known as category theory. From the perspective of a Haskell programmer, however, it is best to think of a monad as an abstract datatype of actions. Haskell's do expressions provide a convenient syntax for writing monadic expressions.

Instances of Monad should satisfy the following:

Left identity

return a >>= k = k a

Right identity

m >>= return = m

Associativity

m >>= (\x -> k x >>= h) = (m >>= k) >>= h

Furthermore, the Monad and Applicative operations should relate as follows:

pure = return

m1 <*> m2 = m1 >>= (x1 -> m2 >>= (x2 -> return (x1 x2)))

The above laws imply:

fmap f xs  =  xs >>= return . f

(>>) = (*>)

and that pure and (<*>) satisfy the applicative functor laws.

The instances of Monad for lists, Maybe and IO defined in the Prelude satisfy these laws.

Minimal complete definition

(>>=)

class IxApplicative m => IxMonad (m :: k -> k -> Type -> Type)

Minimal complete definition

ibind